// src/app/model/agence.model.ts
export class Agence {
    licenceExpirationDate: Date;
    state: String;
    _id:  string;
    title:  String;
    phone: Number;
    email:  String;
    address: String;
    taxRegistrationNum:String;
    taxRegistrationDate:Date;

  postalCode:String;
  region: String;
  cin: String;
  cinDate: Date;

    




}